Update existing bindings
Can compile
